// libs/modules/sps-website-builder/relations/footers-to-widgets/frontend/component/variants/sps-lite/select-right/src/lib/interface.ts
import { IRelation } from "@sps/sps-website-builder-relations-footers-to-widgets-contracts";
import { IRelation as IRelationExtended } from "@sps/sps-website-builder-relations-footers-to-widgets-contracts-extended";

export const variant = "select-right" as const;

export interface IComponentBase {
  showSkeletons?: boolean;
  isServer: boolean;
}

export interface IComponentProps extends IComponentBase {
  variant: typeof variant;
  footerId?: string;
  data?: IRelation;
}

export interface IComponentPropsExtended extends IComponentProps {
  data?: IRelationExtended;
}